fix unloading world bug

+++ src/main/java/com/alpsbte/plotsystem/core/system/plot/PlotHandler.java
@@ -75,24 +75,21 @@ public class PlotHandler {
     }
 
     public static boolean abandonPlot(Plot plot) {
-        if (plot.getWorld().isWorldGenerated() && plot.getWorld().loadWorld()) {
-
-            try {
-                if(plot.getPlotOwner().playInVoid){
-                    for (Player player : plot.getWorld().getBukkitWorld().getPlayers()) {
-                        player.teleport(Utils.getSpawnLocation());
-                    }
-
-                    if (!plot.getWorld().deleteWorld()) {
-                        Bukkit.getLogger().log(Level.SEVERE, "Failed to delete plot world with the ID " + plot.getID() + "!");
-                        return false;
-                    }
+        try {
+            if (plot.getPlotOwner().playInVoid && plot.getWorld().isWorldGenerated() && plot.getWorld().loadWorld()) {
+                for (Player player : plot.getWorld().getBukkitWorld().getPlayers()) {
+                    player.teleport(Utils.getSpawnLocation());
                 }
-            } catch (SQLException throwables) {
-                Bukkit.getLogger().log(Level.SEVERE, "Failed to delete plot world with the ID " + plot.getID() + "!");
-                return false;
-            }
 
+                if (!plot.getWorld().deleteWorld()) {
+                    Bukkit.getLogger().log(Level.SEVERE, "Failed to delete plot world with the ID " + plot.getID() + "!");
+                    return false;
+                }
+
+            }
+        } catch (SQLException throwables) {
+            Bukkit.getLogger().log(Level.SEVERE, "Failed to delete plot world with the ID " + plot.getID() + "!");
+            return false;
         }
 
         try {
